refactor(daemon): re-home the github hook coords, webchat chunking, and fleet upgrade
Three mechanical moves that put code where its subject already lives:

- `src/daemon/github-hook-coords.ts` → `src/github/hook-coords.ts`.
- `WEBCHAT_CHUNK_BYTES` + `chunkText` out of `src/daemon/text.ts` into
  `src/webchat/chunk.ts`; `text.ts` keeps only the `formatErr*` helpers.
- the fleet restart/upgrade cluster off `Daemon` into
  `src/lifecycle/fleet-upgrade.ts` as `FleetUpgradeCoordinator`, with the
  three fields it exclusively owns (`lifecycleInFlight`,
  `fleetUpgradeInFlight`, `fleetExitStarted`) and a narrow host literal.

No delegates left behind: the config-apply host and the CP client deps host
call `this.fleetUpgrade.*` directly, and the test pokes moved to the
coordinator instance.
